added .yml file

+++ app/views.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request, url_for, redirect, flash, g, session
 from app.database import Database, User
 from app import app
-from flask_login import LoginManager, login_required, logout_user, login_user
+from flask_login import LoginManager, login_required, logout_user, login_user, current_user
 
 
 login_manager = LoginManager()
@@ -98,12 +98,16 @@ def logout():
 def landing(username):
     db = Database()
     rows = db.all_business_names()
+    my_rows = db.my_businesses(username)
     names=[]
+    my_names=[]
     for row in rows:
         names.append(row['business_name'])
+    for row in my_rows:
+        my_names.append(row['business_name'])
     if len(rows) == 0:
         flash('No results found')
-    return render_template('landing.html', username=username, names=names)
+    return render_template('landing.html', username=username, names=names, businesses=my_names)
 
 
 @app.route('/search_business', methods=['GET'])
@@ -112,15 +116,21 @@ def search_business():
     if request.method == 'GET':
         keyword_type=request.args.get('keyword_type', None)
         keyword=request.args.get('keyword', None)
+        username=current_user.get_id()
 
         db = Database()
         rows = db.search_business_names(keyword_type, keyword)
+        my_rows = db.my_businesses(username)
+        my_names=[]
+        for row in my_rows:
+            my_names.append(row['business_name'])
         if len(rows) == 0:
             flash('No results found')
             return render_template('landing.html')
         
         else:
-            return render_template('landing.html', names=rows)
+            return render_template('landing.html', names=rows, businesses=my_names)
+
 
 @app.route('/business/<string:business_name>', methods=['GET'])
 def business(business_name):
@@ -128,14 +138,66 @@ def business(business_name):
         db = Database()
         rows = db.search_business('business_name', business_name)
         business = rows.pop(0)
-        return render_template('business.html', business=business)
+        if business['created_by'] == current_user.get_id():
+            return redirect(url_for('my_business', business_name=business_name))
+        else:
+            return redirect(url_for('other_business', business_name=business_name))
+
+
+@app.route('/mbusiness/<string:business_name>', methods=['GET'])
+def my_business(business_name):
+    if request.method == 'GET':
+        db = Database()
+        rows = db.search_business('business_name', business_name)
+        business = rows.pop(0)
+        return render_template('mbusiness.html', business=business)
+
+
+@app.route('/obusiness/<string:business_name>', methods=['GET'])
+def other_business(business_name):
+    if request.method == 'GET':
+        db = Database
+        rows = db.search_business('business_name', business_name)
+        business = rows.pop(0)
+        return render_template('obusiness.html', business=business)
 
 
 @app.route('/delete', methods=['POST'])
 @login_required
 def delete_business():
+    session.pop('_flashes', None)
     business_name = request.form['business_name']
     db=Database()
-    db.delete_business(business_name)
-    flash('Successfully deleted business')
+    row = db.search_business('business_name', business_name)
+    if len(row) == 1:    
+        db.delete_business(business_name)
+        flash('Successfully deleted business')
+    else:
+        flash('Names do not match!')
     return render_template('business.html', business=None)
+
+
+@app.route('/update_business/<string:business_name>', methods = ['GET', 'POST'])
+@login_required
+def update_business(business_name):
+    if request.method == 'POST':
+        name = request.form['name']
+        category = request.form['category']
+        description = request.form['description']
+        location = request.form['location']
+
+        db = Database()
+        rows = db.search_business('business_name', name)
+        if len(rows) == 1 and rows[0]['created_by'] == current_user.get_id():
+            db.update_business(business_name, name, category, description, location)
+            flash('Details successfully updated')
+            return render_template('update_business.html')
+    else:
+        session.pop('_flashes', None)
+        business_name = request.args.get('business_name', None)
+        db = Database()
+        business = db.search_business('business_name', business_name)
+        return render_template('update_business.html', business_name="business_name", business=business)
+
+
+
